add sequential mode test

+++ scripts/test.sh
@@ -12,10 +12,18 @@ then
 fi
 rm out.mp4
 
+./minimp4_x86 -s vectors/foreman.264 out.mp4
+if ! cmp ./out.mp4 vectors/out_sequential_ref.mp4 >/dev/null 2>&1
+then
+    echo sequential test failed
+    exit 1
+fi
+rm out.mp4
+
 ./minimp4_x86 vectors/foreman_slices.264 out.mp4
 if ! cmp ./out.mp4 vectors/out_slices_ref.mp4 >/dev/null 2>&1
 then
-    echo test failed
+    echo slices test failed
     exit 1
 fi
 rm out.mp4
@@ -28,10 +36,18 @@ then
 fi
 rm out.mp4
 
+qemu-arm ./minimp4_arm_gcc -s vectors/foreman.264 out.mp4
+if ! cmp ./out.mp4 vectors/out_sequential_ref.mp4 >/dev/null 2>&1
+then
+    echo sequential test failed
+    exit 1
+fi
+rm out.mp4
+
 qemu-arm ./minimp4_arm_gcc vectors/foreman_slices.264 out.mp4
 if ! cmp ./out.mp4 vectors/out_slices_ref.mp4 >/dev/null 2>&1
 then
-    echo test failed
+    echo slices test failed
     exit 1
 fi
 rm out.mp4
